requirements.txt
Scrapy
scrapyd
scrapyd-client
google-cloud-storage